feat: 공통 네비바의 버튼영역과 프로필 영역 컴포넌트로 구현
