+++ src/components/layout/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './ProfileMenu';
+export * from './RenderLink';
